test: merge test files
